+++ src/components/Shop.js
@@ -1,5 +1,5 @@
-import React from "react";
-import Item from "./Item";
+import React from 'react';
+import Item from './Item';
 
 export default function Shop(props) {
   return (
